+++ src/main/java/DukeException.java
@@ -1,0 +1,10 @@
+public class DukeException extends Exception {
+    public DukeException(String message) {
+        super(message);
+    }
+
+    public DukeException(String message, Throwable cause) {
+        super(message, cause);
+    }
+    
+}
