+++ text-ui-test/input.txt
@@ -1,0 +1,5 @@
+todo borrow book
+deadline return book /by Sunday
+event project meeting /from Mon 2pm /to 4pm
+list
+bye
